vindler_backend/requirements.txt
aioredis==1.3.1
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
boto3==1.16.53
botocore==1.19.53
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
msgpack==1.0.2
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
whitenoise==5.2.0
zope.interface==5.2.0
